blank slate
My current approach was taking too much time, I have to reimplement
every combinator from scala.util.parsing.combinator.Parsers. But the
only reason I could not use Parsers direction is because that API
expects a stream and we have a tree instead. So the new plan is: convert
the tree to a stream, and use Parsers.