strip whitespace
I was planning to do this with my new Reader.filter and Reader.map, but
it turned out to be a bad idea, so I modified the input NodeSeq instead.

The reason this was a bad idea is that the XmlOpen and XmlClose tokens
are holding a NodeSeq representing themselves and their child nodes. For
this reason, filtering out tokens representing unwanted children would
not be sufficient, I would also need to modify those NodeSeqs anyway, so
I might as well choose the solution which only requires modify the
NodeSeq.