simplify whitespace trimming
turns out I don't actually care about preserving any of the outer
whitespace even if it's not strictly indentation. For example, if I want
to talk about the following XML snippet:

  <foo> : <bar><baz/></bar></foo>

I'd like to be able to rewrite the above in multiple lines:

  <foo>
    :
    <bar>
      <baz/>
    </bar>
  </foo>

But then the spaces before and after the colon get swallowed by the
indentation. So I don't think they should have been considered
significant to begin with.